code/tenant/requirements.txt
cassandra-driver
pandas
kafka-python
pyspark
